reading notes
Common issues may be resolved by identifying symptoms and putting measures in place to address them. Typical issues include a lack of power, physical damage, an unsuccessful boot, an application that won't start, and faulty peripherals. Let's take a closer look at a few of them.

A typical start is impossible
When your computer won't boot up, think like a detective and look for clues. What has changed since the last time I checked? Is there any new hardware or software that you've installed? Any wires that may have fallen loose have been moved? Did the operating system restart on its own after the update was applied?
After a recent OS upgrade, the machine may not start up properly at first (that is, slowly or with errors). You may be able to repair it by restarting your computer. You may also use the System Restore tool in Windows to go back to a restore point from the previous day if that doesn't work.
A recovery or troubleshooting mode may be offered if the operating system fails to start at all. If that's an option, take it. If that doesn't work, try booting from the operating system's removable installation disc and selecting Repair to enter the Windows Recovery Environment (RE). Windows 10's Settings app (Update & Security > Recovery > Advanced Startup > Restart Now) may also get you there.

Unable to download or launch a certain software
Hardware incompatibilities, particularly with the display adapter, are typically to blame for problems with a certain programme. The most recent games sometimes need the use of high-performance display adapters and the most recent drivers. The application's installation software should be able to tell you if there are any issues with your computer's hardware. If it doesn't and you start experiencing issues after launching the programme, check online for any known hardware compatibility issues. Try obtaining the most recent driver version from the adapter manufacturer's website to update your video card (display adapter) driver.
The next step is to determine whether a patch or update is available for download from the application's manufacturer's website. A bug in your own system, such as a small incompatibility between a programme and its operating system or some hardware, is likely the cause of any problems you're encountering. Installing the programme again may cure your issue, since this is more common than you may expect.
Is there no joy in sight? Check to verify whether the programme needs a certain version of Windows. A newer version of the software may be available even though it was created for an earlier version of Windows. If it doesn't work, you may try running it in Compatibility Mode, which emulates a previous version of Windows. It's easy to use: Right-click the shortcut for the programme (or its executable file) and choose Troubleshoot compatibility from the menu. To let Windows make an educated guess as to the proper settings, click the Try suggested settings button. To test for compatibility with an earlier version of Windows, use the Compatibility tab in the program's Properties box.

Apps are up and running, although they sometimes freeze.

Many processes might be running at the same time on today's operating systems, especially those on mobile devices. Processes are scheduled to use processor time so that it seems as though many applications are operating at the same time. The problem is that some of these software programmes might malfunction, causing them to become unusable. It is possible to compel them to leave in this situation:
Windows: Using ctrl+alt+del, you can go to the Task Manager, where you can right-click on the errant programme and choose End task to end its life.
MacOS: Force Quit may be found in the Apple menu system.
Linux: To see the PIDs that the Linux kernel has allocated to currently executing processes, use the ps command. To put an end to the process, use the following command with the PID: Killing a process using the SIGKILL signal, -9, may put an end to bad behaviour.
An Android-powered phone: To force stop an app, open the Settings app, pick Apps, select the app from the list, and touch Force Stop.
This is iOS 11.x:  Swipe the app card up and off the screen by pressing the left edge of the iPhone screen and moving your finger to the centre of the screen.